Moving libproc calls in System.Diagnostics.Process over to use shims
