+++ src/Common/src/Interop/OSX/Interop.Libraries.cs
@@ -7,7 +7,6 @@ internal static partial class Interop
     {
         internal const string CoreFoundationLibrary = "/System/Library/Frameworks/CoreFoundation.framework/CoreFoundation";
         internal const string CoreServicesLibrary   = "/System/Library/Frameworks/CoreServices.framework/CoreServices";
-        internal const string libproc = "libproc";
         internal const string LibSystemKernel = "/usr/lib/system/libsystem_kernel";
         internal const string LibCurl = "libcurl";             // Curl HTTP client library
     }
