+++ src/Native/Common/pal_config.h.in
@@ -13,6 +13,7 @@
 #cmakedefine01 HAVE_POSIX_ADVISE
 #cmakedefine01 PRIORITY_REQUIRES_INT_WHO
 #cmakedefine01 HAVE_IN6_U
+#cmakedefine01 HAVE_LIBPROC
 
 // Mac OS X has stat64, but it is deprecated since plain stat now
 // provides the same 64-bit aware struct when targeting OS X > 10.5
