+++ src/Native/System.Security.Cryptography.Native/pal_evp.h
@@ -48,7 +48,6 @@ Direct shim to EVP_DigestFinal_ex.
 */
 extern "C" int32_t EvpDigestFinalEx(EVP_MD_CTX* ctx, unsigned char* md, uint32_t* s);
 
-
 /*
 Function:
 EvpMdSize
