+++ src/Native/System.Security.Cryptography.Native/pal_evp_cipher.h
@@ -14,11 +14,7 @@ Implemented by:
 
 Returns new EVP_CIPHER_CTX on success, nullptr on failure.
 */
-extern "C" EVP_CIPHER_CTX* EvpCipherCreate(
-    const EVP_CIPHER* type, 
-    unsigned char* key, 
-    unsigned char* iv, 
-    int32_t enc);
+extern "C" EVP_CIPHER_CTX* EvpCipherCreate(const EVP_CIPHER* type, unsigned char* key, unsigned char* iv, int32_t enc);
 
 /*
 Cleans up and deletes an EVP_CIPHER_CTX instance created by EvpCipherCreate.
@@ -55,12 +51,8 @@ EvpCipherUpdate
 
 Direct shim to EVP_CipherUpdate.
 */
-extern "C" int32_t EvpCipherUpdate(
-    EVP_CIPHER_CTX* ctx,
-    unsigned char* out,
-    int32_t* outl,
-    unsigned char* in,
-    int32_t inl);
+extern "C" int32_t
+EvpCipherUpdate(EVP_CIPHER_CTX* ctx, unsigned char* out, int32_t* outl, unsigned char* in, int32_t inl);
 
 /*
 Function:
@@ -68,10 +60,7 @@ EvpCipherFinalEx
 
 Direct shim to EVP_CipherFinal_ex.
 */
-extern "C" int32_t EvpCipherFinalEx(
-    EVP_CIPHER_CTX* ctx,
-    unsigned char* outm,
-    int32_t* outl);
+extern "C" int32_t EvpCipherFinalEx(EVP_CIPHER_CTX* ctx, unsigned char* outm, int32_t* outl);
 
 /*
 Function:
